♻️ Refactor: seperate filter from contens
